도메인 변경
회원 도메인 추가 및 스터디 도메인 위치 변경
* 롬복을 사용
* 스터디와 회원은 다대일 연관 매핑
