--- src/main/java/com/example/testjava/Study.java
@@ -1,43 +1,0 @@
-package com.example.testjava;
-
-public class Study {
-
-  private StudyStatus status;
-
-  private int limit;
-
-  private String name;
-
-  public Study(int limit, String name) {
-    this.limit = limit;
-    this.name = name;
-  }
-
-  public Study(int limit) {
-    if (limit < 0) {
-      throw new IllegalArgumentException("limit은 0보다 커야 한다.");
-    }
-    this.limit = limit;
-  }
-
-  public StudyStatus getStatus() {
-    return status;
-  }
-
-  public int getLimit() {
-    return limit;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  @Override
-  public String toString() {
-    return "Study{" +
-            "status=" + status +
-            ", limit=" + limit +
-            ", name='" + name + '\'' +
-            '}';
-  }
-}
