+++ src/main/java/com/example/testjava/domain/Member.java
@@ -1,0 +1,19 @@
+package com.example.testjava.domain;
+
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter @Setter @NoArgsConstructor
+public class Member {
+
+  @Id @GeneratedValue
+  private Long id;
+
+  private String email;
+}
